+++ app/page.tsx
@@ -4,7 +4,9 @@
 
 "use client";
 
+import Image from "next/image";
 import Link from "next/link";
+import { GAME_BRAND } from "../lib/lore";
 import {
   motion,
   useScroll,